azfunction/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
pymysql > 1.0
pathlib > 1.0
dynatrace-opentelemetry-azure-functions 
opentelemetry-instrumentation-pymysql
opentelemetry-api
opentelemetry-sdk